Remove old admin
