+++ shell/client/styles/_admin.scss
@@ -1,250 +1,3 @@
-#admin-settings {
-  margin: 16px;
-  .nav {
-    display: inline-block;
-    padding-left: 5px;
-    padding-bottom: 5px;
-    >li {
-      float: left;
-      position: relative;
-      display: block;
-      >a {
-        padding: 10px;
-        border-radius: 4px;
-        &:hover {
-          background-color: lighten($sandstorm-purple-color, 40%);
-        }
-      }
-      &.active >a {
-        color: white;
-        background-color: $sandstorm-purple-color;
-      }
-    }
-  }
-  h4 {
-    margin-bottom: 0;
-  }
-  .details {
-    font-size: 70%;
-    display: block;
-    margin-bottom: 8px;
-  }
-  .forwork-note {
-    margin-top: 0;
-    font-size: 70%;
-  }
-  .alert {
-    padding: 15px;
-    margin-bottom: 20px;
-    border: 1px solid transparent;
-    border-radius: 4px;
-    -webkit-transition: all 2s ease-in-out;
-    -moz-transition: all 2s ease-in-out;
-    -ms-transition: all 2s ease-in-out;
-    -o-transition: all 2s ease-in-out;
-    transition: all 2s ease-in-out;
-  }
-  .alert-danger {
-    color: #a94442;
-    background-color: #f2dede;
-    border-color: #ebccd1;
-  }
-  .alert-success {
-    color: #3c763d;
-    background-color: #dff0d8;
-    border-color: #d6e9c6;
-  }
-  .fade-out {
-    opacity: .5;
-  }
-  .reset-reason {
-    margin: 10px;
-  }
-}
-
-.settings-tabs {
-  font-size: 16px;
-}
-
-#admin-settings-save {
-  @extend %button-base;
-  @extend %button-primary;
-  margin-top: 20px;
-  margin-bottom: 20px;
-}
-
-#admin-settings-send-toggle {
-  @extend %button-base;
-  @extend %button-secondary;
-  margin-left: 10px;
-}
-
-#admin-settings-send-test {
-  @extend %button-base;
-  @extend %button-secondary;
-}
-
-#admin-settings-form {
-  input[type="text"], input[type="email"], input[type="password"], input[type="number"] {
-    width: 100%;
-  }
-
-  textarea {
-    display: block;
-    width: 100%;
-  }
-}
-
-#settings-content {
-  width: 600px;
-}
-
-table.admin-table {
-  font-size: 14px;
-  border-collapse: collapse;
-  thead td {
-    font-weight: 600;
-    border-bottom: 1px solid #aaa;
-  }
-  td {
-    padding: 0px 5px 5px;
-    &:first-child {
-      padding-left: 16px;
-    }
-    a.disable-cap {
-      color: red;
-      cursor: pointer;
-    }
-  }
-  tr {
-    height: 32px;
-    border-bottom: 1px solid #ddd;
-    &.revoked-cap {
-      text-decoration: line-through;
-    }
-  }
-}
-
-#key-note {
-  width: 300px;
-}
-
-input.autoSelect {
-  font-family: monospace;
-  width: 450px;
-}
-
-#invite {
-  input, select {
-    font: inherit;
-    font-size: inherit;
-    width: 80%;
-  }
-  input#key-quota {
-    width: 40%;
-    margin-bottom: 8px;
-  }
-  textarea {
-    vertical-align: text-top;
-    width: 80%;
-    height: 200px;
-    font: inherit;
-    font-size: inherit;
-  }
-  button {
-    @extend %button-base;
-    @extend %button-primary;
-    font-family: inherit;
-    font-size: 90%;
-    padding: 6px;
-    line-height: 1.33;
-    &:disabled {
-      background-color: #888;
-    }
-  }
-}
-
-#stats {
-  .report-stats-yesno-box {
-    background-color: #D8BFD8;
-    padding: 16px;
-    text-align: right;
-    p {
-      text-align: left;
-      margin-bottom: 16px;
-    }
-    button {
-      margin-right: 5px;
-
-      &.yes {
-        @extend %button-base;
-        @extend %button-primary;
-        padding: 8px 16px;
-      }
-      &.no {
-        background: none;
-        border: none;
-        color: black;
-        text-decoration: underline;
-        font-weight: normal;
-        font-size: 85%;
-      }
-    }
-  }
-  .report-stats {
-    img {
-      width: 16px;
-    }
-    .check-mark {
-      color: green;
-      transition: all 2s ease-in-out;
-      &.fade-out {
-        opacity: .2;
-      }
-    }
-  }
-  >button#regenerateStatsToken {
-    @extend %button-base;
-    @extend %button-primary;
-  }
-
-  td {
-    text-align: center;
-    vertical-align: middle;
-    padding: 8px;
-    border: 1px solid #aaa;
-  }
-  table {
-    border-collapse: collapse;
-    width: 100%;
-    &.apps {
-      tr:nth-child(2) {
-        font-size: 11px;
-      }
-    }
-  }
-}
-button#offer-ipnetwork, button#offer-ipinterface {
-  @extend %button-base;
-  @extend %button-primary;
-  font-size: 14px;
-  min-height: 32px;
-}
-
-.websocket-broken-warning, .wildcard-host-warning {
-  min-height: 3em;
-  width: 100%;
-  padding: 0.5em;
-  margin-bottom: 0.5em;
-  background-color: red;
-  color: white;
-
-  a {
-    color: white;
-    text-decoration: underline;
-  }
-}
-
 .feature-key-data {
   background-color: #ffffff;
   padding: 42px;
@@ -331,17 +84,5 @@ form.feature-key-form, form.feature-key-modify-form {
   @extend %standard-form;
   textarea {
     min-height: 60px;
-  }
-}
-
-.login-providers {
-  label {
-    display: block;
-  }
-}
-
-.legacy-email-config {
-  label {
-    display: block;
   }
 }
