+++ shell/imports/client/personalization.js
@@ -1,0 +1,3 @@
+const DEFAULT_SIGNUP_DIALOG = "You've been invited to join this Sandstorm server!";
+
+export { DEFAULT_SIGNUP_DIALOG };
